+++ app/src/main/java/fr/isen/shazamphoto/database/ToIdentifyMonumentDAO.java
@@ -55,7 +55,7 @@ public class ToIdentifyMonumentDAO extends DAOBase {
             String args2[] = {String.valueOf(c.getLong(2))};
             c = mDb.query(DatabaseHandler.LOCALIZATION_TABLE_NAME, DatabaseHandler.LOCALIZATION_ALL_COLUMNS, DatabaseHandler.LOCALIZATION_KEY + " = ?", args2, "", "", "");
             if(c.moveToFirst()) {
-                monument.setLocalization(new Localization(c.getLong(0), c.getInt(1), c.getInt(2)));
+                monument.setLocalization(new Localization(c.getLong(0), c.getDouble(1), c.getDouble(2)));
             }
         }
         return monument;
@@ -71,7 +71,7 @@ public class ToIdentifyMonumentDAO extends DAOBase {
             String args2[] = {String.valueOf(c.getLong(2))};
             c = mDb.query(DatabaseHandler.LOCALIZATION_TABLE_NAME, DatabaseHandler.LOCALIZATION_ALL_COLUMNS, DatabaseHandler.LOCALIZATION_KEY + " = ?", args2, "", "", "");
             if(c.moveToFirst()) {
-                monument.setLocalization(new Localization(c.getLong(0), c.getInt(1), c.getInt(2)));
+                monument.setLocalization(new Localization(c.getLong(0), c.getDouble(1), c.getDouble(2)));
             }
             monuments.add(monument);
         }
